refactor(OrderValidator): OrderTableDao가 아닌 Repository를 이용하도록 변경
