+++ src/main/java/kitchenpos/order/domain/OrderValidator.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.domain;
 
 import java.util.List;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.domain.OrderTable;
 import kitchenpos.menu.domain.MenuRepository;
+import kitchenpos.ordertable.domain.OrderTable;
+import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -13,11 +13,11 @@ public class OrderValidator {
     private static final String EMPTY_TABLE = "이용중이지 않은 테이블에서는 주문 할 수 없습니다.";
 
     private final MenuRepository menuRepository;
-    private final OrderTableDao orderTableDao;
+    private final OrderTableRepository orderTableRepository;
 
-    public OrderValidator(MenuRepository menuRepository, OrderTableDao orderTableDao) {
+    public OrderValidator(MenuRepository menuRepository, OrderTableRepository orderTableRepository) {
         this.menuRepository = menuRepository;
-        this.orderTableDao = orderTableDao;
+        this.orderTableRepository = orderTableRepository;
     }
 
     public void validateMenu(Order order) {
@@ -28,7 +28,7 @@ public class OrderValidator {
     }
 
     public void validateTable(Order order) {
-        OrderTable table = orderTableDao.findById(order.getOrderTableId())
+        OrderTable table = orderTableRepository.findById(order.getOrderTableId())
                 .orElseThrow(() -> new IllegalArgumentException(NOT_EXIST_TABLE));
 
         if (table.isEmpty()) {
